requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.1
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
Babel==2.9.1
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
distlib==0.3.7
distro==1.9.0
exceptiongroup==1.2.0
filelock==3.13.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
html5lib==1.1
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-metadata==7.0.1
lxml==5.1.0
magic-filter==1.0.12
multidict==6.0.4
openai==1.10.0
packaging==23.2
pefile==2023.2.7
platformdirs==3.11.0
pydantic==2.5.3
pydantic_core==2.14.6
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.12
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32-ctypes==0.2.2
redis==5.0.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
tqdm==4.66.1
typing_extensions==4.9.0
validators==0.22.0
virtualenv==20.24.6
webencodings==0.5.1
yarl==1.9.4
zipp==3.17.0
